Small refactor (move entity method test to entity-test.js)
