+++ lib/entity.js
@@ -53,12 +53,7 @@ class Entity {
             ...opts,
         };
 
-        let { error } = validateEntityData();
-
-        if (!error) {
-            ({ error } = validateMethod(options.method));
-        }
-
+        const { error } = validate();
         if (error) {
             return Promise.reject(error);
         }
@@ -102,6 +97,16 @@ class Entity {
             }
 
             return _this;
+        }
+
+        function validate() {
+            let { error: err } = validateEntityData();
+
+            if (!err) {
+                ({ error: err } = validateMethod(options.method));
+            }
+
+            return { error: err };
         }
 
         function validateEntityData() {
