+++ lib/serializers/datastore.js
@@ -4,49 +4,19 @@ const is = require('is');
 const arrify = require('arrify');
 
 function toDatastore(entity, options = {}) {
-    const excludeFromIndexes = [...entity.excludeFromIndexes] || [];
-
-    let isArray;
-    let isObject;
-    let propConfig;
-
     // For now the more robust excudeFromIndexes Array declaration
     // has an issue with Arrays ("Exclude from indexes cannot be set on a list value")
     // and cannot be used for now
     // See issue: https://github.com/googleapis/nodejs-datastore/issues/14
 
-    const data = Object.keys(entity.entityData).reduce((acc, key) => {
-        if (typeof entity.entityData[key] !== 'undefined') {
-            acc[key] = entity.entityData[key];
+    const data = Object.entries(entity.entityData).reduce((acc, [key, value]) => {
+        if (typeof value !== 'undefined') {
+            acc[key] = value;
         }
         return acc;
     }, {});
 
-    Object.keys(data).forEach((k) => {
-        if (entity.entityData[k] !== null) {
-            propConfig = entity.schema.paths[k];
-
-            isArray = propConfig && (propConfig.type === 'array'
-                || (propConfig.joi && propConfig.joi._type === 'array'));
-
-            isObject = propConfig && (propConfig.type === 'object'
-                || (propConfig.joi && propConfig.joi._type === 'object'));
-
-            if (isArray && propConfig.excludeFromIndexes === true) {
-                // We exclude all the primitives from Array
-                // The format is "entityProp[]"
-                excludeFromIndexes.push(`${k}[]`);
-            } else if (isObject && propConfig.excludeFromIndexes === true) {
-                // For "object" type we automatically set all its properties to excludeFromIndexes: true
-                // which is what most of us expect.
-                Object.keys(entity.entityData[k]).forEach((kk) => {
-                    // We add the embedded property to our Array of excludedFromIndexes
-                    // The format is "entityProp.entityKey"
-                    excludeFromIndexes.push(`${k}.${kk}`);
-                });
-            }
-        }
-    });
+    const excludeFromIndexes = getExcludeFromIndexes();
 
     const datastoreFormat = {
         key: entity.entityKey,
@@ -62,6 +32,46 @@ function toDatastore(entity, options = {}) {
     }
 
     return datastoreFormat;
+
+    // ---------
+
+    function getExcludeFromIndexes() {
+        const excluded = [...entity.excludeFromIndexes] || [];
+        let isArray;
+        let isObject;
+        let propConfig;
+        let propValue;
+
+        Object.keys(data).forEach((prop) => {
+            propValue = entity.entityData[prop];
+            if (propValue === null) {
+                return;
+            }
+            propConfig = entity.schema.paths[prop];
+
+            isArray = propConfig && (propConfig.type === 'array'
+              || (propConfig.joi && propConfig.joi._type === 'array'));
+
+            isObject = propConfig && (propConfig.type === 'object'
+            || (propConfig.joi && propConfig.joi._type === 'object'));
+
+            if (isArray && propConfig.excludeFromIndexes === true) {
+                // We exclude all the primitives from Array
+                // The format is "entityProp[]"
+                excluded.push(`${prop}[]`);
+            } else if (isObject && propConfig.excludeFromIndexes === true) {
+                // For "object" type we automatically set all its properties to excludeFromIndexes: true
+                // which is what most of us expect.
+                Object.keys(propValue).forEach((k) => {
+                    // We add the embedded property to our Array of excludedFromIndexes
+                    // The format is "entityProp.entityKey"
+                    excluded.push(`${prop}.${k}`);
+                });
+            }
+        });
+
+        return excluded;
+    }
 }
 
 function fromDatastore(entity, options = {}) {
